Laravel TestCase and Facades test
